html/obsolete/requirements-for-implementations/other-elements-attributes-and-apis/document-color-01.html

<!DOCTYPE html>
<title>document: fg/bg/link/vlink/alink-color</title>
<link rel="author" title="Ms2ger" href="mailto:ms2ger@gmail.com">
<link rel="help" href="https://html.spec.whatwg.org/multipage/#dom-document-fgcolor">
<link rel="help" href="https://html.spec.whatwg.org/multipage/#dom-body-text">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
test(function() {
  document.fgColor = "green";
  document.bgColor = "green";
  document.linkColor = "green";
  document.vlinkColor = "green";
  document.alinkColor = "green";

  var body = document.documentElement.removeChild(document.body);

  // When there is no body element, the color attributes return an
  // empty string upon getting.
  assert_equals(document.fgColor, "");
  assert_equals(document.bgColor, "");
  assert_equals(document.linkColor, "");
  assert_equals(document.vlinkColor, "");
  assert_equals(document.alinkColor, "");

  // Re-add body and reset color attributes.
  document.body = body;
  document.fgColor = "";
  document.bgColor = "";
  document.linkColor = "";
  document.vlinkColor = "";
  document.alinkColor = "";
}, "getting document color attributes with no body");

test(function() {
  var body = document.documentElement.removeChild(document.body);

  // When there is no body element, setting the color attributes has no effect.
  document.fgColor = "red";
  document.bgColor = "red";
  document.linkColor = "red";
  document.vlinkColor = "red";
  document.alinkColor = "red";
  assert_equals(document.fgColor, "");
  assert_equals(document.bgColor, "");
  assert_equals(document.linkColor, "");
  assert_equals(document.vlinkColor, "");
  assert_equals(document.alinkColor, "");

  document.body = body;
}, "setting document color attributes with no body");

test(function() {
  var doc = document.implementation.createHTMLDocument();
  doc.replaceChild(doc.createElement('test'), doc.documentElement);
  var new_body = doc.createElement('body');
  doc.documentElement.appendChild(new_body);

  doc.fgColor = "red";
  doc.bgColor = "red";
  doc.linkColor = "red";
  doc.vlinkColor = "red";
  doc.alinkColor = "red";

  assert_equals(new_body.attributes.length, 0, 'new_body.attributes.length');
  assert_equals(doc.fgColor, "");
  assert_equals(doc.bgColor, "");
  assert_equals(doc.linkColor, "");
  assert_equals(doc.vlinkColor, "");
  assert_equals(doc.alinkColor, "");
}, "document color attributes when the root element is a test element");
</script>
